Move imageUpload into module
